+++ changelog.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 - zapis zenitovych uhlov pre body stranou
 - aktualizacia formatovacieho suboru GKUv2017.FRT
 - pridane examples
+- pridany command --plx_vysuhl pre pripravu plx suboru na vyskove riesenie polygonu v Kokesi